Update base path
