--- popcornguineapigplugin/src/main/kotlin/com/gabrielbmoro/popcorn/domain/input/PopcornConfiguration.kt
@@ -1,8 +1,0 @@
-package com.gabrielbmoro.popcorn.domain.input
-
-import com.gabrielbmoro.popcorn.domain.rules.PopcornGuineaPigRule
-
-data class PopcornConfiguration(
-    val project: PopcornProject,
-    val rules: List<PopcornGuineaPigRule>
-)
