--- popcornguineapigplugin/src/main/kotlin/com/gabrielbmoro/popcorn/domain/input/PopcornProject.kt
@@ -1,6 +1,0 @@
-package com.gabrielbmoro.popcorn.domain.input
-
-data class PopcornProject(
-    val type: ProjectType,
-    val groupName: String? = null
-)
