--- popcornguineapigplugin/src/main/kotlin/com/gabrielbmoro/popcorn/domain/input/ProjectType.kt
@@ -1,7 +1,0 @@
-package com.gabrielbmoro.popcorn.domain.input
-
-enum class ProjectType {
-    JAVA,
-    KMP,
-    ANDROID
-}
