--- popcornguineapigplugin/src/main/kotlin/com/gabrielbmoro/popcorn/domain/metadata/InternalDependenciesMetadata.kt
@@ -1,6 +1,0 @@
-package com.gabrielbmoro.popcorn.domain.metadata
-
-data class InternalDependenciesMetadata(
-    val group: String?,
-    val moduleName: String
-)
