--- popcornguineapigplugin/src/main/kotlin/com/gabrielbmoro/popcorn/domain/output/CheckResult.kt
@@ -1,9 +1,0 @@
-package com.gabrielbmoro.popcorn.domain.output
-
-internal sealed class CheckResult {
-    object Success : CheckResult()
-
-    data class Failure(
-        val errors: List<ArchitectureViolationError>
-    ) : CheckResult()
-}
