--- popcornguineapigplugin/src/main/kotlin/com/gabrielbmoro/popcorn/domain/usecases/GetRightConfigurationNameUseCase.kt
@@ -1,12 +1,0 @@
-package com.gabrielbmoro.popcorn.domain.usecases
-
-import com.gabrielbmoro.popcorn.domain.input.ProjectType
-
-
-internal class GetRightConfigurationNameUseCase {
-
-    fun execute(projectType: ProjectType) = when (projectType) {
-        ProjectType.KMP -> "commonMainImplementation"
-        else -> "implementation"
-    }
-}
