+++ popcornguineapigplugin/src/main/kotlin/com/github/codandotv/popcorn/domain/input/PopcornProject.kt
@@ -1,0 +1,6 @@
+package com.github.codandotv.popcorn.domain.input
+
+data class PopcornProject(
+    val type: ProjectType,
+    val groupName: String? = null
+)
