+++ popcornguineapigplugin/src/main/kotlin/com/github/codandotv/popcorn/domain/input/ProjectType.kt
@@ -1,0 +1,7 @@
+package com.github.codandotv.popcorn.domain.input
+
+enum class ProjectType {
+    JAVA,
+    KMP,
+    ANDROID
+}
