+++ popcornguineapigplugin/src/main/kotlin/com/github/codandotv/popcorn/domain/metadata/InternalDependenciesMetadata.kt
@@ -1,0 +1,6 @@
+package com.github.codandotv.popcorn.domain.metadata
+
+data class InternalDependenciesMetadata(
+    val group: String?,
+    val moduleName: String
+)
