+++ popcornguineapigplugin/src/main/kotlin/com/github/codandotv/popcorn/domain/rules/NoDependencyRule.kt
@@ -1,0 +1,17 @@
+package com.github.codandotv.popcorn.domain.rules
+
+import com.github.codandotv.popcorn.domain.output.ArchitectureViolationError
+import com.github.codandotv.popcorn.domain.metadata.InternalDependenciesMetadata
+
+class NoDependencyRule : PopcornGuineaPigRule {
+    override fun check(deps: List<InternalDependenciesMetadata>): ArchitectureViolationError? {
+        if (deps.isNotEmpty()) {
+            return ArchitectureViolationError(
+                message = "NoDependencyRule: This module should not have dependencies",
+                rule = this
+            )
+        }
+
+        return null
+    }
+}
