+++ popcornguineapigplugin/src/main/kotlin/com/github/codandotv/popcorn/domain/rules/PopcornGuineaPigRule.kt
@@ -1,0 +1,9 @@
+package com.github.codandotv.popcorn.domain.rules
+
+import com.github.codandotv.popcorn.domain.output.ArchitectureViolationError
+import com.github.codandotv.popcorn.domain.metadata.InternalDependenciesMetadata
+
+interface PopcornGuineaPigRule {
+
+    fun check(deps: List<InternalDependenciesMetadata>): ArchitectureViolationError?
+}
