+++ popcornguineapigplugin/src/main/kotlin/com/github/codandotv/popcorn/domain/usecases/CheckArchitectureUseCase.kt
@@ -1,0 +1,33 @@
+package com.github.codandotv.popcorn.domain.usecases
+
+import com.github.codandotv.popcorn.domain.output.ArchitectureViolationError
+import com.github.codandotv.popcorn.domain.output.CheckResult
+import com.github.codandotv.popcorn.domain.metadata.TargetModule
+import com.github.codandotv.popcorn.domain.input.PopcornConfiguration
+
+internal class CheckArchitectureUseCase {
+
+    fun execute(configuration: PopcornConfiguration, targetModule: TargetModule): CheckResult {
+        val sortedInternalProjectDependencies = targetModule.internalDependencies
+            .distinct()
+            .sortedBy {
+                it.moduleName
+            }
+
+        val errors = mutableListOf<com.github.codandotv.popcorn.domain.output.ArchitectureViolationError>()
+
+        configuration.rules.forEach { rule ->
+            rule.check(sortedInternalProjectDependencies)?.let { error ->
+                errors.add(error)
+            }
+        }
+
+        return if (errors.isEmpty()) {
+            CheckResult.Success
+        } else {
+            CheckResult.Failure(
+                errors = errors
+            )
+        }
+    }
+}
