+++ popcornguineapigplugin/src/test/kotlin/com/github/codandotv/popcorn/domain/usecases/FakePopcornConfigurations.kt
@@ -1,0 +1,12 @@
+package com.github.codandotv.popcorn.domain.usecases
+
+import com.github.codandotv.popcorn.domain.input.PopcornConfiguration
+import com.github.codandotv.popcorn.domain.input.PopcornProject
+import com.github.codandotv.popcorn.domain.input.ProjectType
+
+val fakePopcornConfiguration = PopcornConfiguration(
+    project = PopcornProject(
+        type = ProjectType.KMP
+    ),
+    rules = emptyList()
+)
