--- sample/sample-data/src/main/java/com/gabrielbmoro/sample/Data.kt
@@ -1,5 +1,0 @@
-package com.gabrielbmoro.sample
-
-class Data {
-
-}
