+++ sample/sample-data/src/main/java/com/github/codandotv/sample/Data.kt
@@ -1,0 +1,5 @@
+package com.github.codandotv.sample
+
+class Data {
+
+}
