--- sample/sample-domain/src/main/java/com/gabrielbmoro/sample/Domain.kt
@@ -1,5 +1,0 @@
-package com.gabrielbmoro.sample
-
-class Domain {
-
-}
