+++ sample/sample-domain/src/main/java/com/github/codandotv/sample/Domain.kt
@@ -1,0 +1,5 @@
+package com.github.codandotv.sample
+
+class Domain {
+
+}
