--- sample/sample-main/src/main/java/com/gabrielbmoro/sample/Main.kt
@@ -1,5 +1,0 @@
-package com.gabrielbmoro.sample
-
-fun main() {
-    println("Test")
-}
