+++ sample/sample-main/src/main/java/com/github/codandotv/sample/Main.kt
@@ -1,0 +1,5 @@
+package com.github.codandotv.sample
+
+fun main() {
+    println("Test")
+}
