--- sample/sample-presentation/src/main/java/com/gabrielbmoro/sample/Presentation.kt
@@ -1,5 +1,0 @@
-package com.gabrielbmoro.sample
-
-class Presentation {
-
-}
