+++ sample/sample-presentation/src/main/java/com/github/codandotv/sample/Presentation.kt
@@ -1,0 +1,5 @@
+package com.github.codandotv.sample
+
+class Presentation {
+
+}
